+++ packages/mobile/src/components/JoinCommunity/JoinCommunity.component.tsx
@@ -10,6 +10,7 @@ import { JoinCommunityProps } from './JoinCommunity.types'
 import { getInvitationCodes } from '@quiet/state-manager'
 
 import { Splash } from '../Splash/Splash.component'
+import { InvitationData } from '@quiet/types'
 
 export const JoinCommunity: FC<JoinCommunityProps> = ({
   joinCommunityAction,
@@ -39,8 +40,14 @@ export const JoinCommunity: FC<JoinCommunityProps> = ({
       return
     }
 
-    const submitValue = getInvitationCodes(joinCommunityInput.trim())
-    if (!submitValue?.length) {
+    let submitValue: InvitationData | null = null
+    try {
+      submitValue = getInvitationCodes(joinCommunityInput.trim())
+    } catch (e) {
+      console.warn(`Could not parse invitation code, reason: ${e.message}`)
+    }
+
+    if (!submitValue) {
       setLoading(false)
       setInputError('Please check your invitation code and try again')
       return