+++ packages/mobile/src/components/JoinCommunity/JoinCommunity.types.ts
@@ -1,7 +1,7 @@
-import { InvitationPair } from '@quiet/types'
+import { InvitationData } from '@quiet/types'
 
 export interface JoinCommunityProps {
-  joinCommunityAction: (address: InvitationPair[]) => void
+  joinCommunityAction: (address: InvitationData) => void
   redirectionAction: () => void
   networkCreated: boolean
   invitationCode?: string