+++ packages/mobile/src/tests/splash.screen.test.tsx
@@ -10,6 +10,7 @@ import { ScreenNames } from '../const/ScreenNames.enum'
 import { initActions } from '../store/init/init.slice'
 import { take } from 'typed-redux-saga'
 import { navigationActions } from '../store/navigation/navigation.slice'
+import { validInvitationUrlTestData } from '@quiet/common'
 
 describe('Splash screen', () => {
   let socket: MockedSocket
@@ -22,8 +23,7 @@ describe('Splash screen', () => {
   test('waits for redux store to become ready, before storing invitation code', async () => {
     const { store, root, runSaga } = await prepareStore({}, socket)
 
-    const invitationCode =
-      'QmZoiJNAvCffeEHBjk766nLuKVdkxkAT7wfFJDPPLsbKSE=y7yczmugl2tekami7sbdz5pfaemvx7bahwthrdvcbzw5vex2crsr26qd'
+    const invitationCode = validInvitationUrlTestData.code()
 
     const route: { key: string; name: ScreenNames.SplashScreen; path: string } = {
       key: '',