+++ packages/types/src/network.ts
@@ -7,3 +7,8 @@ export type InvitationPair = {
   peerId: string
   onionAddress: string
 }
+
+export type InvitationData = {
+  pairs: InvitationPair[]
+  psk: string
+}