+++ packages/integration-tests/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.2-alpha.2](https://github.com/TryQuiet/quiet/compare/integration-tests@2.0.2-alpha.1...integration-tests@2.0.2-alpha.2) (2023-11-14)
+
+**Note:** Version bump only for package integration-tests
+
+
+
+
+
+## [2.0.2-alpha.1](https://github.com/TryQuiet/quiet/compare/integration-tests@2.0.2-alpha.0...integration-tests@2.0.2-alpha.1) (2023-11-13)
+
+**Note:** Version bump only for package integration-tests
+
+
+
+
+
 ## [2.0.2-alpha.0](https://github.com/TryQuiet/quiet/compare/integration-tests@2.0.1-alpha.6...integration-tests@2.0.2-alpha.0) (2023-10-26)
 
 **Note:** Version bump only for package integration-tests